refactor(sentry): split the selector's second look and decision reporting out
Adding the bounded second look pushed `sentry-autofix-select.mjs` to 694 lines,
past the repo's 600-line soft cap. Split rather than append, per the checklist.

`sentry-autofix-second-look.mjs` owns the "the first pass found nothing off a
FULL page — what is past the ceiling?" pass: its ceilings, its halved family
budgets, and the cost arithmetic that sizes them against the 25-minute timeout.
It returns RAW decisions, so the classification stays in one place.

`sentry-autofix-decisions.mjs` owns that classification — the deferral wording
and the decision -> report split both passes run through, which is what keeps a
second-look family deferral from being reported as a fix_scope skip.

Pure move plus the import rewiring; the selector re-exports the second look's
ceilings, since the suite and the workflow read the selection contract from it.
